ekf2: move output predictor to output_predictor/
