+++ src/modules/ekf2/EKF/output_predictor/output_predictor.h
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2022 PX4. All rights reserved.
+ *   Copyright (c) 2022-2024 PX4. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -36,8 +36,7 @@
 
 #include <matrix/math.hpp>
 
-#include "common.h"
-#include "RingBuffer.h"
+#include "../RingBuffer.h"
 
 #include <lib/geo/geo.h>
